Use the same core code for both Aquarium versions
Instead of having two nearly identical copies of the code, it's
simpler to differentiate the VR and non-VR versions with just a
simple configuration file.

This switches to more similar canvas resizing logic in both VR and
non-VR versions of Aquarium. The canvas is resized in non-VR mode only
if fitWindow is set.
