+++ aquarium-vr/aquarium-vr.html
@@ -110,11 +110,12 @@ CANVAS {
 <script type="text/javascript" src="../jquery-ui-1.8.2.custom/js/jquery-ui-1.8.2.custom.min.js"></script>
 <script type="text/javascript" src="../khronos/webgl-debug.js"></script>
 <script type="text/javascript" src="../tdl/base.js"></script>
-<script type="text/javascript" src="../aquarium/aquarium-common.js"></script>
 <script type="text/javascript" src="utils/vr_text_renderer.js"></script>
 <script type="text/javascript" src="utils/gamepad.js"></script>
 <script type="text/javascript" src="utils/ui.js"></script>
-<script type="text/javascript" src="aquarium-vr.js"></script>
+<script type="text/javascript" src="aquarium-config.js"></script>
+<script type="text/javascript" src="../aquarium/aquarium-common.js"></script>
+<script type="text/javascript" src="../aquarium/aquarium.js"></script>
 </head>
 <body>
 <div id="info"><a href="http://threedlibrary.googlecode.com" target="_blank">tdl.js</a> - aquarium</div>
