+++ aquarium/aquarium-config.js
@@ -1,0 +1,6 @@
+"use strict";
+
+var g_aquariumConfig = {
+    aquariumRoot: '',
+    enableVR: false
+};
