+++ aquarium/aquarium-control-panel.js
@@ -1,13 +1,9 @@
 tdl.require('tdl.clock');
-tdl.require('tdl.fast');
 tdl.require('tdl.io');
 tdl.require('tdl.log');
-tdl.require('tdl.math');
 tdl.require('tdl.sync');
 
 // globals
-var math;                 // the math lib.
-var fast;                 // the fast math lib.
 var g_setSettingElements = [];
 var g_numSettingElements = {};
 var g_inactiveSeconds = 0;
@@ -38,9 +34,6 @@ function checkActivity() {
  * Initializes stuff.
  */
 function initialize() {
-  math = tdl.math;
-  fast = tdl.fast;
-
   initializeCommon();
 
   setInterval(checkActivity, 1000);
