+++ aquarium/aquarium.html
@@ -110,6 +110,7 @@ CANVAS {
 <script type="text/javascript" src="../jquery-ui-1.8.2.custom/js/jquery-ui-1.8.2.custom.min.js"></script>
 <script type="text/javascript" src="../khronos/webgl-debug.js"></script>
 <script type="text/javascript" src="../tdl/base.js"></script>
+<script type="text/javascript" src="aquarium-config.js"></script>
 <script type="text/javascript" src="aquarium-common.js"></script>
 <script type="text/javascript" src="aquarium.js"></script>
 </head>
